requirements.txt
paramiko>=2.7.2
Jinja2>=2.11.3
PyYAML>=5.4.1
click>=7.1.2